ISBN converted to string. it seems some ISBN13 starts with 0
yeah interesting right? it's only one arabic book I found with this pattern but
it seems to be following te ISBN13 format rules
https://www.goodreads.com/book/show/39785977
ISBN: 0663546946043

So we can't have it as bigint as 0 will be removed when saving. and as now it's
a string we can use rails validations for length.

oh and also I forgot to make it unique to the user. I was able to add same book
twice all this time.